Forbid Sentinel conversion to bool
Going with a strict implementation as early as possible.
This restriction can always be relaxed at a later time.

A truth test on a sentinel implies that the sentinel is mixed in
with other types during the test.
In nearly all cases this is incorrect and could result in
unexpected behavior.